requirements.txt
# Web Framework
streamlit>=1.43.2

# Data Processing
pandas>=2.2.3
numpy>=1.23.0
openpyxl>=3.1.5
xlrd>=2.0.1

# AI/LLM Providers
anthropic>=0.49.0
openai>=1.66.3

# Database Engines
duckdb>=0.9.0
databricks-sql-connector>=2.9.0

# Healthcare Data Standards
hl7>=0.4.5
fhir-r4b>=0.0.1

# Templates and Configuration
pyyaml>=6.0.1
jinja2>=3.1.0

# Validation and Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Performance and Monitoring
prometheus-client>=0.17.0
structlog>=22.0.0

# Utilities
plotly>=6.0.0
trafilatura>=2.0.0
twilio>=9.5.0
requests>=2.27.0
gitpython>=3.1.0

# Development Tools
black>=22.0.0
mypy>=1.0.0